Задание 4. Разделила код на модули
